Update lab9.cpp
compiles on my computer and works fine. Using Dev C++. Only thing is that the search doesn't work for 3 digit GPA's for some reason (i.e. 3.79). Otherwise, everything works.

I was thinking about implementing a type checker in case someone inputs a string for an int (I've done a few times by accident), but that could be a whole number can of worms.